refactor: move modal tests
