+++ frontend/svelte/src/tests/lib/modals/neurons/NewFolloweeModal.spec.ts
@@ -5,29 +5,29 @@
 import { Topic } from "@dfinity/nns";
 import { fireEvent } from "@testing-library/dom";
 import { render } from "@testing-library/svelte";
-import NewFolloweeModal from "../../../lib/modals/neurons/NewFolloweeModal.svelte";
+import NewFolloweeModal from "../../../../lib/modals/neurons/NewFolloweeModal.svelte";
 import {
   addFollowee,
   removeFollowee,
-} from "../../../lib/services/neurons.services";
-import { authStore } from "../../../lib/stores/auth.store";
-import { knownNeuronsStore } from "../../../lib/stores/knownNeurons.store";
-import { mockAuthStoreSubscribe } from "../../mocks/auth.store.mock";
-import en from "../../mocks/i18n.mock";
+} from "../../../../lib/services/neurons.services";
+import { authStore } from "../../../../lib/stores/auth.store";
+import { knownNeuronsStore } from "../../../../lib/stores/knownNeurons.store";
+import { mockAuthStoreSubscribe } from "../../../mocks/auth.store.mock";
+import en from "../../../mocks/i18n.mock";
 import {
   mockFullNeuron,
   mockKnownNeuron,
   mockNeuron,
-} from "../../mocks/neurons.mock";
+} from "../../../mocks/neurons.mock";
 
-jest.mock("../../../lib/services/neurons.services", () => {
+jest.mock("../../../../lib/services/neurons.services", () => {
   return {
     addFollowee: jest.fn().mockResolvedValue(undefined),
     removeFollowee: jest.fn().mockResolvedValue(undefined),
   };
 });
 
-jest.mock("../../../lib/services/knownNeurons.services", () => {
+jest.mock("../../../../lib/services/knownNeurons.services", () => {
   return {
     listKnownNeurons: jest.fn(),
   };
